move around code to make it easier to understand
